// src/gql/graphql.ts
export type Maybe<T> = T | null
// ...
export type Scalars = {
	ID: { input: string; output: string }
	String: { input: string; output: string }
	Boolean: { input: boolean; output: boolean }
	Int: { input: number; output: number }
	Float: { input: number; output: number }
	Bool: { input: any; output: any }
	Date: { input: any; output: any }
	DateTime: { input: any; output: any }
	/** The `JSON` scalar type represents JSON values as specified by [ECMA-404](https://ecma-international.org/publications-and-standards/standards/ecma-404/). */
	JSON: { input: any; output: any }
}
// ...
export type ArticlePage = IData &
	_IContent &
	_IPage & {
		__typename?: 'ArticlePage'
		/** @deprecated Use `_link` field instead */
		_children?: Maybe<QueryRef>
		_deleted?: Maybe<Scalars['Bool']['output']>
		_fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		_id?: Maybe<Scalars['String']['output']>
		_link?: Maybe<QueryRef>
		_metadata?: Maybe<IContentMetadata>
		_modified?: Maybe<Scalars['Date']['output']>
		_score?: Maybe<Scalars['Float']['output']>
		articleAuthors?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		articleBody?: Maybe<RichText>
		articleHeroImage?: Maybe<ContentReference>
		articleSeoSettings?: Maybe<PageSeoSettingsProperty>
		articleSummary?: Maybe<RichText>
		articleTitle?: Maybe<Scalars['String']['output']>
	}
// ...
export type ContentMetadata = IContentMetadata & {
	__typename?: 'ContentMetadata'
	created?: Maybe<Scalars['DateTime']['output']>
	displayName?: Maybe<Scalars['String']['output']>
	key?: Maybe<Scalars['String']['output']>
	lastModified?: Maybe<Scalars['DateTime']['output']>
	locale?: Maybe<Scalars['String']['output']>
	published?: Maybe<Scalars['DateTime']['output']>
	status?: Maybe<Scalars['String']['output']>
	types?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	url?: Maybe<ContentUrl>
	version?: Maybe<Scalars['String']['output']>
}

export type ContentMetadatadisplayNameArgs = {
	highlight?: InputMaybe<HighlightOptions>
}

export type ContentReference = {
	__typename?: 'ContentReference'
	key?: Maybe<Scalars['String']['output']>
	url?: Maybe<ContentUrl>
}
// ...
export type ContentUrl = {
	__typename?: 'ContentUrl'
	base?: Maybe<Scalars['String']['output']>
	default?: Maybe<Scalars['String']['output']>
	hierarchical?: Maybe<Scalars['String']['output']>
	internal?: Maybe<Scalars['String']['output']>
	type?: Maybe<Scalars['String']['output']>
}
// ...
export type IContentMetadata = {
	created?: Maybe<Scalars['DateTime']['output']>
	displayName?: Maybe<Scalars['String']['output']>
	key?: Maybe<Scalars['String']['output']>
	lastModified?: Maybe<Scalars['DateTime']['output']>
	locale?: Maybe<Scalars['String']['output']>
	published?: Maybe<Scalars['DateTime']['output']>
	status?: Maybe<Scalars['String']['output']>
	types?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	url?: Maybe<ContentUrl>
	version?: Maybe<Scalars['String']['output']>
}
// ...
export type IData = {
	/** @deprecated Use `_link` field instead */
	_children?: Maybe<QueryRef>
	_deleted?: Maybe<Scalars['Bool']['output']>
	_fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	_id?: Maybe<Scalars['String']['output']>
	_link?: Maybe<QueryRef>
	_modified?: Maybe<Scalars['Date']['output']>
	_score?: Maybe<Scalars['Float']['output']>
}
// ...
export type InstanceMetadata = IContentMetadata &
	IInstanceMetadata & {
		__typename?: 'InstanceMetadata'
		container?: Maybe<Scalars['String']['output']>
		created?: Maybe<Scalars['DateTime']['output']>
		createdBy?: Maybe<Scalars['String']['output']>
		displayName?: Maybe<Scalars['String']['output']>
		expired?: Maybe<Scalars['DateTime']['output']>
		key?: Maybe<Scalars['String']['output']>
		lastModified?: Maybe<Scalars['DateTime']['output']>
		lastModifiedBy?: Maybe<Scalars['String']['output']>
		locale?: Maybe<Scalars['String']['output']>
		locales?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		owner?: Maybe<Scalars['String']['output']>
		path?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		published?: Maybe<Scalars['DateTime']['output']>
		routeSegment?: Maybe<Scalars['String']['output']>
		status?: Maybe<Scalars['String']['output']>
		types?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		url?: Maybe<ContentUrl>
		version?: Maybe<Scalars['String']['output']>
	}

export type InstanceMetadatadisplayNameArgs = {
	highlight?: InputMaybe<HighlightOptions>
}

export type ItemMetadata = IContentMetadata &
	IItemMetadata & {
		__typename?: 'ItemMetadata'
		created?: Maybe<Scalars['DateTime']['output']>
		displayName?: Maybe<Scalars['String']['output']>
		displayOption?: Maybe<Scalars['String']['output']>
		key?: Maybe<Scalars['String']['output']>
		lastModified?: Maybe<Scalars['DateTime']['output']>
		locale?: Maybe<Scalars['String']['output']>
		published?: Maybe<Scalars['DateTime']['output']>
		status?: Maybe<Scalars['String']['output']>
		types?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		url?: Maybe<ContentUrl>
		version?: Maybe<Scalars['String']['output']>
	}
// ...
export type MediaMetadata = IContentMetadata &
	IInstanceMetadata &
	IMediaMetadata & {
		__typename?: 'MediaMetadata'
		container?: Maybe<Scalars['String']['output']>
		content?: Maybe<Scalars['String']['output']>
		created?: Maybe<Scalars['DateTime']['output']>
		createdBy?: Maybe<Scalars['String']['output']>
		displayName?: Maybe<Scalars['String']['output']>
		expired?: Maybe<Scalars['DateTime']['output']>
		key?: Maybe<Scalars['String']['output']>
		lastModified?: Maybe<Scalars['DateTime']['output']>
		lastModifiedBy?: Maybe<Scalars['String']['output']>
		locale?: Maybe<Scalars['String']['output']>
		locales?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		mimeType?: Maybe<Scalars['String']['output']>
		owner?: Maybe<Scalars['String']['output']>
		path?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		published?: Maybe<Scalars['DateTime']['output']>
		routeSegment?: Maybe<Scalars['String']['output']>
		status?: Maybe<Scalars['String']['output']>
		thumbnail?: Maybe<Scalars['String']['output']>
		types?: Maybe<Array<Maybe<Scalars['String']['output']>>>
		url?: Maybe<ContentUrl>
		version?: Maybe<Scalars['String']['output']>
	}
// ...
export type QueryRef = {
	__typename?: 'QueryRef'
	ArticleGroupPage?: Maybe<ArticleGroupPageOutput>
	ArticleListElement?: Maybe<ArticleListElementOutput>
	ArticlePage?: Maybe<ArticlePageOutput>
	BannerBlock?: Maybe<BannerBlockOutput>
	BlankExperience?: Maybe<BlankExperienceOutput>
	BlankSection?: Maybe<BlankSectionOutput>
	ButtonBlock?: Maybe<ButtonBlockOutput>
	CTAElement?: Maybe<CTAElementOutput>
	CardBlock?: Maybe<CardBlockOutput>
	Data?: Maybe<DataOutput>
	DefaultImage?: Maybe<DefaultImageOutput>
	DefaultVideo?: Maybe<DefaultVideoOutput>
	Dictionary?: Maybe<DictionaryOutput>
	DictionaryItem?: Maybe<DictionaryItemOutput>
	GenericMedia?: Maybe<GenericMediaOutput>
	HeaderBlock?: Maybe<HeaderBlockOutput>
	HeadingElement?: Maybe<HeadingElementOutput>
	ImageElement?: Maybe<ImageElementOutput>
	ImageMedia?: Maybe<ImageMediaOutput>
	MegaMenuGroupBlock?: Maybe<MegaMenuGroupBlockOutput>
	NavigationMenuBlock?: Maybe<NavigationMenuBlockOutput>
	OfficeLocation?: Maybe<OfficeLocationOutput>
	PageSeoSettings?: Maybe<PageSeoSettingsOutput>
	ParagraphElement?: Maybe<ParagraphElementOutput>
	SysContentFolder?: Maybe<SysContentFolderOutput>
	TestimonialElement?: Maybe<TestimonialElementOutput>
	VideoMedia?: Maybe<VideoMediaOutput>
	WebsiteFooter?: Maybe<WebsiteFooterOutput>
	_Component?: Maybe<_ComponentOutput>
	_Content?: Maybe<_ContentOutput>
	_Element?: Maybe<_ElementOutput>
	_Experience?: Maybe<_ExperienceOutput>
	_Folder?: Maybe<_FolderOutput>
	_Image?: Maybe<_ImageOutput>
	_Media?: Maybe<_MediaOutput>
	_Page?: Maybe<_PageOutput>
	_Section?: Maybe<_SectionOutput>
	_Video?: Maybe<_VideoOutput>
}
// ...
export type RichText = {
	__typename?: 'RichText'
	html?: Maybe<Scalars['String']['output']>
	json?: Maybe<Scalars['JSON']['output']>
}
// ...
export type _IContent = {
	/** @deprecated Use `_link` field instead */
	_children?: Maybe<QueryRef>
	_deleted?: Maybe<Scalars['Bool']['output']>
	_fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	_id?: Maybe<Scalars['String']['output']>
	_link?: Maybe<QueryRef>
	_metadata?: Maybe<IContentMetadata>
	_modified?: Maybe<Scalars['Date']['output']>
	_score?: Maybe<Scalars['Float']['output']>
}
// ...
export type _IPage = {
	/** @deprecated Use `_link` field instead */
	_children?: Maybe<QueryRef>
	_deleted?: Maybe<Scalars['Bool']['output']>
	_fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>
	_id?: Maybe<Scalars['String']['output']>
	_link?: Maybe<QueryRef>
	_metadata?: Maybe<IContentMetadata>
	_modified?: Maybe<Scalars['Date']['output']>
	_score?: Maybe<Scalars['Float']['output']>
}
// ...
export type ReferenceDataFragment = {
	__typename?: 'ContentReference'
	key?: string | null
	url?:
		| ({ __typename?: 'ContentUrl' } & {
				' $fragmentRefs'?: { LinkDataFragment: LinkDataFragment }
		  })
		| null
} & { ' $fragmentName'?: 'ReferenceDataFragment' }
// ...
export type ArticlePageDataFragment = {
	__typename?: 'ArticlePage'
	articleAuthors?: Array<string | null> | null
	articleTitle?: string | null
	metadata?:
		| { __typename?: 'ContentMetadata'; published?: any | null }
		| { __typename?: 'InstanceMetadata'; published?: any | null }
		| { __typename?: 'ItemMetadata'; published?: any | null }
		| { __typename?: 'MediaMetadata'; published?: any | null }
		| null
	articleHeroImage?:
		| ({ __typename?: 'ContentReference' } & {
				' $fragmentRefs'?: { ReferenceDataFragment: ReferenceDataFragment }
		  })
		| null
	articleBody?: { __typename?: 'RichText'; json?: any | null } | null
  heroImageUrl1?: string | null
	heroImageUrl2?: string | null
	imagePosition?: string | null
} & { ' $fragmentName'?: 'ArticlePageDataFragment' }
// ...
export const ArticlePageDataFragmentDoc = {
	kind: 'Document',
	definitions: [
		{
			kind: 'FragmentDefinition',
			name: { kind: 'Name', value: 'ArticlePageData' },
			typeCondition: { kind: 'NamedType', name: { kind: 'Name', value: 'ArticlePage' } },
			selectionSet: {
				kind: 'SelectionSet',
				selections: [
					{
						kind: 'Field',
						alias: { kind: 'Name', value: 'metadata' },
						name: { kind: 'Name', value: '_metadata' },
						selectionSet: {
							kind: 'SelectionSet',
							selections: [
								{ kind: 'Field', name: { kind: 'Name', value: 'published' } },
							],
						},
					},
					{
						kind: 'Field',
						name: { kind: 'Name', value: 'articleHeroImage' },
						selectionSet: {
							kind: 'SelectionSet',
							selections: [
								{
									kind: 'FragmentSpread',
									name: { kind: 'Name', value: 'ReferenceData' },
								},
							],
						},
					},
					{ kind: 'Field', name: { kind: 'Name', value: 'articleAuthors' } },
					{ kind: 'Field', name: { kind: 'Name', value: 'articleTitle' } },
					{
						kind: 'Field',
						name: { kind: 'Name', value: 'articleBody' },
						selectionSet: {
							kind: 'SelectionSet',
							selections: [{ kind: 'Field', name: { kind: 'Name', value: 'json' } }],
						},
					},
				],
			},
		},
		{
			kind: 'FragmentDefinition',
			name: { kind: 'Name', value: 'LinkData' },
			typeCondition: { kind: 'NamedType', name: { kind: 'Name', value: 'ContentUrl' } },
			selectionSet: {
				kind: 'SelectionSet',
				selections: [
					{ kind: 'Field', name: { kind: 'Name', value: 'base' } },
					{ kind: 'Field', name: { kind: 'Name', value: 'hierarchical' } },
					{ kind: 'Field', name: { kind: 'Name', value: 'default' } },
				],
			},
		},
		{
			kind: 'FragmentDefinition',
			name: { kind: 'Name', value: 'ReferenceData' },
			typeCondition: { kind: 'NamedType', name: { kind: 'Name', value: 'ContentReference' } },
			selectionSet: {
				kind: 'SelectionSet',
				selections: [
					{ kind: 'Field', name: { kind: 'Name', value: 'key' } },
					{
						kind: 'Field',
						name: { kind: 'Name', value: 'url' },
						selectionSet: {
							kind: 'SelectionSet',
							selections: [
								{
									kind: 'FragmentSpread',
									name: { kind: 'Name', value: 'LinkData' },
								},
							],
						},
					},
				],
			},
		},
	],
} as unknown as DocumentNode<ArticlePageDataFragment, unknown>